requirements.txt
opencv-python
click
termcolor
pillow
numpy